Rename and adjust imports for library and books pages
Reorganized file structure by renaming and relocating `library` and `books-list` pages, simplifying their paths. Updated all associated imports to reflect the new structure for better maintainability.
